refactor: add PollViewPage
